services/aggregator/requirements.txt
paho-mqtt==1.6.1
pydantic==2.8.2
orjson==3.10.7
